Update to main.rb
