+++ app.rb
@@ -4,40 +4,12 @@ require './book'
 require './student'
 require './classroom'
 require './rental'
-require './print_script'
 
-class App < PrintScript
+class App
   def initialize
     super
     @books = []
     @persons = []
     @rentals = []
-  end
-
-  def list_all_books
-    puts 'Database is empty! Add a book.' if @books.empty?
-    @books.each { |book| puts "[Book] Title: #{book.title}, Author: #{book.author}" }
-  end
-
-  def list_all_persons
-    puts 'Database is empty! Add a person.' if @persons.empty?
-    @persons.each do |person|
-      puts "[#{person.class.name}] Name: #{person.name}, Age: #{person.age}, id: #{person.id}"
-    end
-  end
-
-  def list_all_rentals
-    puts 'To see rentals enter the person ID: '
-    id = gets.chomp.to_i
-
-    puts 'Rented Books: '
-    @rentals.each do |rental|
-      if rental.person.id == id
-        puts "Person: #{rental.person.name} Date: #{rental.date}, Book: '#{rental.book.title}' by #{rental.book.author}"
-      else
-        puts
-        puts 'No record were found for the given ID'
-      end
-    end
   end
 end
